RL - Added Payment page - Updated dashboard to reflect CreditsAvailable

+++ rooferlocator.com.Web/Models/Member/MemberSummaryViewModel.cs
@@ -5,6 +5,7 @@ using rooferlocator.com.Users;
 using CreditsHero.Subscribers.Dtos;
 using rooferlocator.com.Common;
 using System.Collections.Generic;
+using System;
 
 namespace rooferlocator.com.Web.Models.Member
 {
@@ -14,7 +15,7 @@ namespace rooferlocator.com.Web.Models.Member
         /// Not required for single-tenant applications.
         /// </summary>
         public int SubscriberInquiriesCount { get; set; }
-        public int CreditsRemainingCount { get; set; }
+        public Decimal CreditsRemainingCount { get; set; }
         public CreditsHero.Subscribers.Dtos.SubscribersInquiriesDto SubscriberInquiries { get; set; }
         public List<CreditsHero.Subscribers.Dtos.SubscribersRequestDto> SubscriberRequests { get; set; }
         public MemberQuotesViewModel SubscriberQuotes { get; set; }
